refactor: move render types to `frui::render`
