+++ crates/frui_core/src/api/implementers/inherited.rs
@@ -1,11 +1,8 @@
 use std::any::TypeId;
 
 use crate::{
-    api::{
-        contexts::render_ctx::{paint_ctx::PaintCtxOS, LayoutCtxOS},
-        IntoWidgetPtr, WidgetPtr,
-    },
-    prelude::{Canvas, Constraints, Offset, Size},
+    api::{IntoWidgetPtr, WidgetPtr},
+    render::*,
 };
 
 use super::{InheritedWidgetOS, WidgetDerive};
