+++ crates/frui_core/src/api/implementers/render.rs
@@ -1,16 +1,8 @@
 use crate::{
-    api::{
-        contexts::{
-            build_ctx::_BuildCtx,
-            render_ctx::{
-                paint_ctx::{PaintCtx, PaintCtxOS},
-                LayoutCtxOS,
-            },
-            Context,
-        },
-        IntoWidgetPtr, WidgetPtr,
-    },
-    prelude::{BuildCtx, Canvas, Constraints, LayoutCtx, Offset, Size},
+    api::{contexts::build_ctx::_BuildCtx, IntoWidgetPtr, WidgetPtr},
+    macro_exports::Context,
+    prelude::BuildCtx,
+    render::*,
 };
 
 use super::{RenderWidgetOS, WidgetDerive};
