+++ crates/frui_core/src/api/implementers/view.rs
@@ -2,12 +2,11 @@ use crate::{
     api::{
         contexts::{
             build_ctx::{BuildCtx, _BuildCtx},
-            render_ctx::{paint_ctx::PaintCtxOS, LayoutCtxOS},
             Context,
         },
         IntoWidgetPtr, WidgetPtr,
     },
-    prelude::{Canvas, Constraints, Offset, Size},
+    render::*,
 };
 
 use super::{ViewWidgetOS, WidgetDerive};
