+++ crates/frui_widgets/src/boxes.rs
@@ -1,4 +1,5 @@
 use frui::prelude::*;
+use frui::render::*;
 
 use crate::BoxLayoutData;
