+++ crates/frui_widgets/src/flex/stack.rs
@@ -2,6 +2,7 @@ use crate::alignment::{Alignment, AlignmentDirectional};
 use crate::{AlignmentGeometry, BoxLayoutData, TextDirection, WidgetList};
 
 use frui::prelude::*;
+use frui::render::*;
 
 pub enum StackFit {
     Loose,
