+++ crates/frui_widgets/src/transform.rs
@@ -1,4 +1,5 @@
 use frui::prelude::*;
+use frui::render::*;
 
 #[derive(RenderWidget)]
 pub struct Transform<W: Widget>(pub Affine, pub W);
