+++ examples/flex.rs
@@ -1,6 +1,7 @@
 #![feature(type_alias_impl_trait)]
 
 use frui::prelude::*;
+use frui::render::*;
 
 mod misc;
