+++ examples/pointer_events.rs
@@ -3,6 +3,7 @@
 #![feature(type_alias_impl_trait)]
 
 use frui::prelude::*;
+use frui::render::*;
 
 #[derive(ViewWidget)]
 struct App;
