+++ examples/transform.rs
@@ -6,6 +6,7 @@
 #![feature(type_alias_impl_trait)]
 
 use frui::prelude::*;
+use frui::render::*;
 
 mod counter;
